Added a MinigameStageFactory
